@@ -3,39 +3,46 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "bootstrap": "^4.3.1",
-    "dotenv": "^6.2.0",
-    "html-react-parser": "0.0.4",
-    "ipfs-api": "^26.1.2",
-    "moment": "^2.24.0",
-    "nodemon": "^2.0.4",
-    "react": "^16.6.3",
-    "react-dom": "^16.6.3",
-    "react-router": "^3.2.1",
-    "react-router-dom": "^4.3.1",
-    "react-scripts": "2.1.1",
-    "truffle-hdwallet-provider": "^1.0.0-web3one.5",
-    "web3": "^1.0.0-beta.37"
+    "bootstrap": "^5.2.3",
+    "dotenv": "^16.0.3",
+    "html-react-parser": "^3.0.9",
+    "ipfs-http-client": "^60.0.0",
+    "moment": "^2.29.4",
+    "nodemon": "^2.0.22",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-router": "^6.8.1",
+    "react-router-dom": "^6.8.1",
+    "react-scripts": "^5.0.1",
+    "truffle-hdwallet-provider": "^1.0.17",
+    "web3": "^1.8.2"
   },
   "engines": {
-    "node": "9.4.0",
-    "npm": "5.7.1",
-    "yarn": "1.3.2"
+    "node": ">=16.0.0",
+    "npm": ">=8.0.0"
   },
   "scripts": {
     "start": "react-scripts start",
     "build": "react-scripts build",
     "test": "react-scripts test",
-    "eject": "react-scripts eject",
-    "postinstall": "npm run start"
+    "eject": "react-scripts eject"
   },
   "eslintConfig": {
-    "extends": "react-app"
+    "extends": [
+      "react-app",
+      "react-app/jest"
+    ]
   },
-  "browserslist": [
-    ">0.2%",
-    "not dead",
-    "not ie <= 11",
-    "not op_mini all"
-  ]
+  "browserslist": {
+    "production": [
+      ">0.2%",
+      "not dead",
+      "not op_mini all"
+    ],
+    "development": [
+      "last 1 chrome version",
+      "last 1 firefox version",
+      "last 1 safari version"
+    ]
+  }
 }
