@@ -1,35 +1,31 @@
-import React, { Component } from "react";
-
-import {Router, Route, browserHistory, Redirect} from "react-router";
-
-import Home from './components/Home'
-import StudentHome from './components/StudentHome'
-import NewRecord from './components/NewRecord'
+import React, { Component } from 'react';
+import { BrowserRouter as Router, Route, Routes, Navigate } from 'react-router-dom';
+import Home from './components/Home';
+import StudentHome from './components/StudentHome';
+import NewRecord from './components/NewRecord';
 import ViewRecord from './components/ViewRecord';
-import Educational from './components/EducationalDetails/Educational'
-import EducationHome from './components/EducationHome'
-import RecordDetails from './components/RecordDetails'
-
-import "./App.css";
+import Educational from './components/EducationalDetails/Educational';
+import EducationHome from './components/EducationHome';
+import RecordDetails from './components/RecordDetails';
+import './App.css';
 
 class App extends Component {
-
   render() {
-      return (
-        <Router history={browserHistory}>
-            <Redirect from="/" to="/home" />
-            <Route>
-              <Route path="Student" component={StudentHome}/>
-              <Route path="newRecord" component={NewRecord}/>
-              <Route path = "viewrecord/:recordId" component = {ViewRecord}/>
-              <Route path = "home" component = {Home}></Route>
-              <Route path = "Institute" component = {EducationHome}></Route>
-              <Route path = "recordData/Education/:recordId/" component = {Educational}/>
-              <Route path = "educationUpdate/:recordId" component = {RecordDetails}></Route>
-
-            </Route>
-        </Router>
+    return (
+      <Router>
+        <Routes>
+          <Route path="/" element={<Navigate to="/home" />} />
+          <Route path="/Student" element={<StudentHome />} />
+          <Route path="/newRecord" element={<NewRecord />} />
+          <Route path="/viewrecord/:recordId" element={<ViewRecord />} />
+          <Route path="/home" element={<Home />} />
+          <Route path="/Institute" element={<EducationHome />} />
+          <Route path="/recordData/Education/:recordId" element={<Educational />} />
+          <Route path="/educationUpdate/:recordId" element={<RecordDetails />} />
+        </Routes>
+      </Router>
     );
   }
 }
+
 export default App;
